Optimizing

+++ aws_interface.py
@@ -9,10 +9,10 @@ client = boto3.client('ec2')
 
 def delete_instance_by_ip(ip):
     print "DELETING " + ip 
-    client_description_response = client.describe_instances()
+    client_description_response = client.describe_instances(Filters = {'Name':'private-ip-address', 'Values':[ip]})
     reservations = client_description_response['Reservations']
     for reservation in reservations:
-        if ip == reservation['Instances'][0]['PrivateIpAddress']:
+        if 'PrivateIpAddress' in reservation['Instances'][0].keys() and ip == reservation['Instances'][0]['PrivateIpAddress']:
             instance_id = reservation['Instances'][0]['InstanceId']
             client.terminate_instances(InstanceIds = [instance_id])
